Mineria_Aprendizaje/requirements.txt
awsebcli==3.20.5
botocore==1.29.81
cement==2.8.2
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.3
contourpy==1.1.0
cycler==0.11.0
docx2pdf==0.1.8
et-xmlfile==1.1.0
fonttools==4.40.0
fpdf==1.7.2
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pathspec==0.10.1
pdfkit==1.0.0
Pillow==9.5.0
pluggy==1.0.0
protobuf==3.20.3
pyparsing==3.1.0
pypiwin32==223
pytest==7.3.1
python-dateutil==2.8.2
python-docx==0.8.11
python-nmap==0.7.1
pytz==2023.3
pywin32==305
PyYAML==5.4.1
requests==2.26.0
scapy==2.5.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.8.5
six==1.14.0
sklearn==0.0.post5
SQLAlchemy==2.0.15
termcolor==1.1.0
threadpoolctl==3.1.0
tk==0.1.0
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.14
wcwidth==0.1.9
